% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_to_text.R
\name{code_to_text}
\alias{code_to_text}
\title{Convert Morse Code symbols to text}
\usage{
code_to_text(string)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}
}
\value{
None (invisible \code{NULL}).
}
\description{
This function converts a character vector of Morse Code symbols and prints
the equivalent text in the console. The function uses the International Morse Code
encoding for 26 letters, 10 numerals, and 15 punctuation marks.
}
\examples{
my_code <- ".... . .-.. .-.. ---"
code_to_text(my_code)
}
