% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse_code.R
\name{morse_code}
\alias{morse_code}
\title{Print Morse Code dictionary}
\usage{
morse_code()
}
\value{
None (invisible \code{NULL})
}
\description{
This function prints the International Morse Code symbols in
the console.
}
\examples{
morse_code()
}
