% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse.R
\name{morse_to_text}
\alias{morse_to_text}
\title{Convert Morse characters back to text}
\usage{
morse_to_text(morse)
}
\arguments{
\item{morse}{String with Morse characters separated by spaces. Must use \code{.}
and \code{-} as dots and bars.}
}
\value{
A string of letters from whatever valid Morse code it can find in the
input. It's fault tolerant - it will ignore bad Morse code.
}
\description{
Convert Morse characters back to text
}
\examples{
text_to_morse_chars("Hello how are you") \%>\% morse_to_text()
# This is fault tolerant of bad Morse code:
morse <-  ".... . ..-.. .-AA.. ---"
morse_to_text(morse)
}
