% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_to_code.R
\name{text_to_code}
\alias{text_to_code}
\title{Convert text to Morse Code symbols}
\usage{
text_to_code(string)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}
}
\value{
None (invisible \code{NULL}).
}
\description{
This function converts a character vector into equivalent
Morse Code symbols (called "dots" and "dashes" or "dits" and "dahs") and
prints them to the console. The function uses the International Morse Code
encoding for 26 letters, 10 numerals, and 15 punctuation marks.
}
\examples{
my_text <- "hello world!"
text_to_code(my_text)
}
