% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse.R
\name{text_to_morse_chars}
\alias{text_to_morse_chars}
\title{Convert ASCII text to morse code}
\usage{
text_to_morse_chars(text)
}
\arguments{
\item{text}{The text to convert. Ignores anything except the letters a-z
(both upper and lower case), numbers, and spaces.}
}
\value{
A character string of Morse code.
}
\description{
Convert ASCII text to morse code
}
\examples{
text_to_morse_chars("SOS SOS SOS")
}
