% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse.R
\name{text_to_morse_matrix}
\alias{text_to_morse_matrix}
\title{Convert text to numeric matrix of Morse code}
\usage{
text_to_morse_matrix(text, line_length = 40)
}
\arguments{
\item{text}{The text to convert. Ignores anything except the letters a-z
(both upper and lower case), numbers, and spaces.}

\item{line_length}{Line length to use (number of columns in the resulting
matrix). Default is 40.}
}
\value{
A matrix with a numeric binary representation of Morse code. Three 1s
in a row represent a bar; a 1 represents a dot; a 0 is a separator. Each
row of Morse code is separated from the others by a row of zeros. This is
mostly useful for plotting and art.
}
\description{
Convert text to numeric matrix of Morse code
}
\examples{
text_to_morse_matrix("SOS SOS SOS")
text_to_morse_matrix("SOS SOS SOS", line_length = 10)
}
