% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse.R
\name{text_to_morse_numeric}
\alias{text_to_morse_numeric}
\title{Converts a text into a binary representation of Morse code}
\usage{
text_to_morse_numeric(text)
}
\arguments{
\item{text}{The text to convert. Ignores anything except the letters a-z
(both upper and lower case), numbers, and spaces.}
}
\value{
A numeric vector with a binary representation of Morse code (with 1 =
dot, 111 = bar, 0 = separator).
}
\description{
Converts a text into a binary representation of Morse code
}
\examples{
text_to_morse_numeric("SOS SOS SOS")
}
