% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse.R
\name{text_to_morse_segments}
\alias{text_to_morse_segments}
\title{Convert text to a data frame with coordinates of Morse code segments}
\usage{
text_to_morse_segments(text, line_length = 40)
}
\arguments{
\item{text}{The text to convert. Ignores anything except the letters a-z
(both upper and lower case), numbers, and spaces.}

\item{line_length}{Line length to use (number of columns in the resulting
matrix). Default is 40.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with segment coordinates of Morse code. This is
mostly useful for plotting and art.
}
\description{
Convert text to a data frame with coordinates of Morse code segments
}
\examples{
library(ggplot2)
text_to_morse_segments("SOS SOS SOS") \%>\%
  ggplot() +
  geom_segment(aes(x = x, xend = xend, y = y, yend = yend), size = 2)

text_to_morse_segments("SOS SOS SOS", line_length = 10)
}
