% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse.R
\name{text_to_morse_sounds}
\alias{text_to_morse_sounds}
\title{Converts text to Morse code sounds}
\usage{
text_to_morse_sounds(
  text,
  pulse_duration = 0.25,
  play = TRUE,
  sampling_freq = 8000,
  carrier_freq = 440
)
}
\arguments{
\item{text}{The text to convert. Ignores anything except the letters a-z
(both upper and lower case), numbers, and spaces.}

\item{pulse_duration}{How long the dot lasts, in seconds. Default is 0.25
seconds. Bars last for three times as long.}

\item{play}{Whether to play the resulting sound immediately. Default is
\code{TRUE}.}

\item{sampling_freq}{The sampling frequency. Default is 8000.}

\item{carrier_freq}{The carrier tone. Deafult is 440 Hz (flat A).}
}
\value{
A soundwave matrix that can be played via \link[audio:play]{audio::play}.
}
\description{
Converts text to Morse code sounds
}
\examples{
\dontrun{
sound <- text_to_morse_sounds("SOS SOS SOS")
audio::play(sound, rate = 8000)
audio::play(sound, rate = 24000)}
}
